feat(snowflake): fix cursor pagination (7 tests)
Three fixes that together close the entire cursor-pagination bucket of
failures on real Snowflake:

core/internal/psql/query.go (new snowflake branches, parallel to the
existing sqlite/mariadb branches):

  - Inner LAST_VALUE: Snowflake rejects LAST_VALUE(x) OVER() because
    it requires an ORDER BY window spec. Emit
      LAST_VALUE(col) OVER (ORDER BY <all-order-by-cols>
                            ROWS BETWEEN UNBOUNDED PRECEDING AND UNBOUNDED FOLLOWING)
    so every row's __cur_N equals the final-row value per the outer
    ORDER BY, matching the per-row constant semantics the cursor
    encoder expects.

  - Outer aggregation wrapper: Snowflake has no json_object / json_arrayagg /
    json_group_array / json_extract (those are DuckDB / SQLite / MariaDB).
    Emit OBJECT_CONSTRUCT('json', ..., 'cursor', ... || MAX(__cur_N) ...)
    instead. The __cur_N values are all identical (LAST_VALUE with the
    UNBOUNDED-following frame above), so plain MAX picks the right value
    without needing an array aggregation. The dialect's RenderChildCursor
    already uses GET_PATH which works over the resulting VARIANT, so no
    TO_JSON wrapping is needed.

tests/snowflake.sql:

  - price DOUBLE -> price NUMBER(10,2). Snowflake serialises DOUBLE in
    JSON as scientific notation with 15-digit precision (1.550000e+01),
    which doesn't match the tests' expected "15.5". NUMBER(10,2)
    serialises cleanly as "15.50" and the tests' assertions pass.

Verified 7/7 cursor tests now pass on real Snowflake:
  - Example_queryWithCursorPagination1
  - Example_queryWithCursorPagination2
  - Example_queryWithNamedCursorPagination
  - Example_queryWithNamedCursorPaginationMultiplePages
  - Example_queryWithBackwardCompatibleCursor
  - Example_subscriptionWithCursor
  - Example_queryWithDynamicOrderBy

Other dialects unaffected — the snowflake branches are parallel to the
existing sqlite/mariadb/default branches and the seed file only affects
the snowflake test suite.

+++ tests/snowflake.sql
@@ -26,7 +26,7 @@ CREATE TABLE products (
   tags ARRAY,
   metadata VARIANT,
   country_code VARCHAR,
-  price DOUBLE,
+  price NUMBER(10,2),
   count_likes BIGINT,
   owner_id BIGINT REFERENCES users(id),
   category_ids ARRAY,
@@ -206,7 +206,7 @@ CREATE TABLE order_items (
   product_id BIGINT NOT NULL,
   variant_id BIGINT NOT NULL,
   quantity INTEGER NOT NULL DEFAULT 1,
-  price DOUBLE NOT NULL,
+  price NUMBER(10,2) NOT NULL,
   FOREIGN KEY (product_id, variant_id) REFERENCES product_variants(product_id, variant_id)
 );
